\name{createFormulaFromDF}
\alias{createFormulaFromDF}
\title{createFormulaFromDF}
\usage{
  createFormulaFromDF(df)
}
\arguments{
  \item{df}{dataframe}
}
\value{
  modelFormula text string of formula that includes
  variables and log variables in df
}
\description{
  Creates text for the 'kitchen sink' formula.
}
\examples{
\dontrun{}
}
\keyword{creation}
\keyword{formula}

