\name{createFullDT}
\alias{createFullDT}
\title{createFullDT}
\usage{
  createFullDT(formulaToUse, localDT)
}
\arguments{
  \item{formulaToUse}{formula}

  \item{localDT}{dataframe of potential input variables to
  model}
}
\value{
  fullDTList list of
}
\description{
  Get log and interaction columns
}
\examples{
\dontrun{}
}
\keyword{transforms}

