\name{createModelEquation}
\alias{createModelEquation}
\title{createModelEquation}
\usage{
  createModelEquation(modelReturn)
}
\arguments{
  \item{modelReturn}{censReg model results}
}
\value{
  combo string
}
\description{
  Create the model equation including coefficients.
}
\examples{
\dontrun{}
}
\keyword{equation}

