\name{generateParamChoices}
\alias{generateParamChoices}
\title{generateParamChoices}
\usage{
  generateParamChoices(predictVariables, modelReturn,
    pathToSave = "", save = FALSE)
}
\arguments{
  \item{predictVariables}{string vector of prediction
  variables}

  \item{modelReturn}{censReg object returned from censReg}

  \item{pathToSave}{string}

  \item{save}{logical}
}
\description{
  Generate a file to be used to narrow down allowed
  parameters in model creation.
}
\examples{
\dontrun{}
}
\keyword{list}
\keyword{model}
\keyword{parameter}

