\name{getMultipleUV}
\alias{getMultipleUV}
\title{getMultipleUV}
\usage{
  getMultipleUV(Site, BeginDate, UVP)
}
\arguments{
  \item{Site}{character, USGS site ID}

  \item{BeginDate}{Date}

  \item{UVP}{character vector}
}
\value{
  UV dataframe
}
\description{
  Get unit value data from multiple parameters.
}
\examples{
\dontrun{}
}
\keyword{data}
\keyword{NWIS}
\keyword{retrieval}

