\name{getPredictVariables}
\alias{getPredictVariables}
\title{getPredictVariables}
\usage{
  getPredictVariables(DTnames)
}
\arguments{
  \item{DTnames}{column names of DT dataframe}
}
\value{
  predictVariables string predict variables based on column
  headers
}
\description{
  Returns a character vector of prediction variables.
}
\examples{
\dontrun{}
}
\keyword{predict}

