\docType{package}
\name{GLRIRegression-package}
\alias{GLRIRegression-package}
\title{GLRI Regression package}
\description{
  \tabular{ll}{ Package: \tab GLRIRegression\cr Type: \tab
  Package\cr Version: \tab 1.0.0\cr Date: \tab
  2013-04-26\cr License: \tab Unlimited for this package,
  dependencies have more restrictive licensing.\cr
  Copyright: \tab This software is in the public domain
  because it contains materials that originally came from
  the United States Geological Survey, an agency of the
  United States Department of Interior. For more
  information, see the official USGS copyright policy at
  http://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
  LazyLoad: \tab yes\cr }
}
\details{
  Collection of functions to simplify data retrieval, model
  creation, and model analysis.
}
\author{
  Laura De Cicco \email{ldecicco@usgs.gov}
}
\keyword{data,}
\keyword{regressions}
\keyword{retrieval,}

