\name{internalStudentRes}
\alias{internalStudentRes}
\title{internalStudentRes}
\usage{
  internalStudentRes(DT, modelCoef, logResponse = FALSE)
}
\arguments{
  \item{DT}{DTframe that includes all response and
  predictor variables}

  \item{modelCoef}{named list of model coefficients}

  \item{logResponse}{logical}
}
\value{
  StRes.all
}
\description{
  Compute internally Studentized residuals
}
\examples{
\dontrun{}
}
\keyword{residuals}
\keyword{studentized}

