\name{longDFResponse}
\alias{longDFResponse}
\title{longDFResponse}
\usage{
  longDFResponse(x, predictVariables, responseVariable)
}
\arguments{
  \item{x}{dataframe in wide format}

  \item{predictVariables}{character vector}

  \item{responseVariable}{character}
}
\value{
  Long dataframe with value, variable, and response columns
}
\description{
  Reshapes a wide dataframe to long.
}
\examples{
\dontrun{}
}
\keyword{long}
\keyword{reshape}

