\name{plotQQTransforms}
\alias{plotQQTransforms}
\title{plotQQTransforms}
\usage{
  plotQQTransforms(localDT, responseVariable,
    printLinear = TRUE, printLog = TRUE,
    printSquare = FALSE, printSQRT = FALSE,
    printEXP = FALSE)
}
\arguments{
  \item{localDT}{dataframe in wide format}

  \item{responseVariable}{character}

  \item{printLinear}{logical}

  \item{printLog}{logical}

  \item{printSquare}{logical}

  \item{printSQRT}{logical}

  \item{printEXP}{logical}
}
\value{
  plot
}
\description{
  Plots Q-Q of predict variables.
}
\examples{
\dontrun{}
}
\keyword{qq}

