\name{resultPlots}
\alias{resultPlots}
\title{resultPlots}
\usage{
  resultPlots(localDT, finalModel, siteINFO)
}
\arguments{
  \item{localDT}{dataframe in wide format}

  \item{finalModel}{censReg model results}

  \item{siteINFO}{dataframe including station name
  (station.nm) and siteID (site.no) (easiestly retrieved
  from dataRetrieval package)}
}
\value{
  plot
}
\description{
  Plots summary plots
}
\examples{
\dontrun{}
}
\keyword{scatterplot}

