\name{saveModelResults}
\alias{saveModelResults}
\title{saveModelResults}
\usage{
  saveModelResults(pathToSave, DT, UV, modelReturn,
    siteINFO)
}
\arguments{
  \item{pathToSave}{full path for saving results}

  \item{DT}{dataframe of potential input variables to
  model}

  \item{UV}{dataframe of unit values}

  \item{modelReturn}{results from censReg}

  \item{siteINFO}{dataframe that includes station.nm and
  site.no}
}
\description{
  Saves critical objects for model analysis
}
\examples{
\dontrun{}
}
\keyword{model}
\keyword{results}
\keyword{save}

