\name{simpleCap}
\alias{simpleCap}
\title{simpleCap}
\usage{
  simpleCap(x)
}
\arguments{
  \item{x}{string}
}
\value{
  retVal corrected string
}
\description{
  Return string with initial capital letters
}
\examples{
exString <- "BAD RIVER"
exString <- simpleCap(exString)
}
\keyword{formatting}
\keyword{string}

