\name{summaryPrintout}
\alias{summaryPrintout}
\title{summaryPrintout}
\usage{
  summaryPrintout(modelReturn, siteINFO,
    saveOutput = FALSE, fileName)
}
\arguments{
  \item{modelReturn}{censReg model results}

  \item{siteINFO}{dataframe including station name
  (station.nm) and siteID (site.no) (easiestly retrieved
  from dataRetrieval package)}

  \item{saveOutput}{logical}

  \item{fileName}{string full path and file name to save
  to}
}
\description{
  Summarize the model results
}
\examples{
\dontrun{}
}
\keyword{summary}
\keyword{text}

